Use Jetty LifeCycle listener to create lock file for healthcheck
Fixes #1

This commit suppresses of-watchdog from creating a lock file because it
can create the lock file before Jetty is actually ready to receive
traffic, causing the function to be labeled ready before it can
successfully be invoked.

Instead, this commit adds a LifeCycle listener to the Jetty instance
that will create the lock file when the Jetty server has started.